+++ lib/typing/converters/string_converter_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestArrayConverter(t *testing.T) {
+func TestArrayConverter_Convert(t *testing.T) {
 	// Array
 	{
 		// Normal arrays
@@ -20,5 +20,13 @@ func TestArrayConverter(t *testing.T) {
 		val, err := ArrayConverter{}.Convert(constants.ToastUnavailableValuePlaceholder)
 		assert.NoError(t, err)
 		assert.Equal(t, `["__debezium_unavailable_value"]`, val)
+	}
+}
+
+func TestIntegerConverter_Convert(t *testing.T) {
+	for _, val := range []any{42, int8(42), int16(42), int32(42), int64(42)} {
+		parsedVal, err := IntegerConverter{}.Convert(val)
+		assert.NoError(t, err)
+		assert.Equal(t, "42", parsedVal)
 	}
 }
